Fix a problem with large-ish numbers

+++ philo/init_sim.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   init_sim.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: slakner <slakner@student.42.fr>            +#+  +:+       +#+        */
+/*   By: stephanie.lakner <stephanie.lakner@stud    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/14 17:02:10 by slakner           #+#    #+#             */
-/*   Updated: 2023/02/03 18:05:23 by slakner          ###   ########.fr       */
+/*   Updated: 2023/02/05 19:14:17 by stephanie.l      ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -24,9 +24,9 @@ t_sim	*init_sim(int num_ph, int time_die, int time_eat, int time_sl)
 	pthread_mutex_init(&(sim->m_dead), NULL);
 	pthread_mutex_init(&(sim->m_full), NULL);
 	pthread_mutex_init(&(sim->m_start), NULL);
-	sim->time_die = time_die * 1000;
-	sim->time_eat = time_eat * 1000;
-	sim->time_sleep = time_sl * 1000;
+	sim->time_die = (long) time_die * 1000;
+	sim->time_eat = (long) time_eat * 1000;
+	sim->time_sleep = (long) time_sl * 1000;
 	sim->sim_has_started = 0;
 	sim->philo_dead = 0;
 	return (sim);
